+++ packages/minimap/src/types.ts
@@ -11,9 +11,12 @@ export type MiniMapProps<NodeData = any> = Omit<HTMLAttributes<SVGSVGElement>, '
   nodeBorderRadius?: number;
   nodeStrokeWidth?: number;
   maskColor?: string;
+  maskStrokeColor?: string;
+  maskStrokeWidth?: number;
   position?: PanelPosition;
   onClick?: (event: MouseEvent, position: XYPosition) => void;
   onNodeClick?: (event: MouseEvent, node: Node<NodeData>) => void;
   pannable?: boolean;
   zoomable?: boolean;
+  ariaLabel?: string | null;
 };